requirements.txt
# Need this because mysql-replication on internal pypi is old.
# Since it is a hit, no fallback to python's pypi happens
--extra-index-url https://pypi.python.org/simple
--extra-index-url https://pypi.python.org/pypi
-e .
MySQL-python==1.2.5
PyMySQL==0.6.7
PyStaticConfiguration==0.9.0
PyYAML==3.11
SQLAlchemy==1.0.12
boto==2.36.0
contextdecorator==0.10.0
data_pipeline==0.4.9
dateglob==0.1
enum34==1.0.4
future==0.14.3
futures==3.0.5
geogrid==1.0.8
ipaddress==1.0.16
iso8601==0.1.10
kafka-python==0.9.5.post5
kazoo==2.2
mccabe==0.5.0
mrjob==0.4.3.post7
-e git://github.com/cccc1999/python-mysql-replication.git#egg=mysql-replication
pep8==1.5.7
pii_generator==0.1.5
ply==3.4
py==1.4.29
pycrypto==2.6.1
pyflakes==0.8.1
pyramid-yelp-conn==0.4.1
pysensu_yelp==0.2.3
pysubnettree==0.23
python-dateutil==2.4.2
pytz==2016.4
requests==2.7.0
send-nsca==0.1.4.1
setproctitle==1.1.8
simplejson==3.6.5
six==1.9.0
-e git://github.com/Yelp/sqlparse.git@4fbab9c395752294225f60393cc0a19d6fb62d54#egg=sqlparse
testing.mysqld
thriftpy==0.1.15
vmprof==0.2.7
wheel==0.24.0
wsgiref==0.1.2
yelp_batch==0.27.1
yelp-cgeom==1.3.1
yelp-clog==2.5.2
yelp-lib==10.0.5
yelp-logging==1.1.2
yelp_kafka>=4.2.0
yelp-servlib==4.3.0
